refactor(data-download): #499 — delete the export-template layer; export every live field of the type
An ExportTemplate persisted exactly four things: a file name, a format, a document type, and an ordered
subset of that type's fields. The first three are per-export parameters that never needed a table. The
fourth was the only real configuration — and it was redundant twice over.

Ordering is already owned by FieldDefinition.DisplayOrder, the axis the operator list renders its dynamic
columns by. ExportColumn.Order was a second, independent axis over the same fields, free to disagree with
what the operator sees. That is the drift class #496 existed to close ("the file must not disagree with the
view"), reintroduced one layer down. And #414 part 2 had already made a freshly created config prefill every
field of the type, so a new template produced byte-identical output to having no template at all — an entity
whose default state is indistinguishable from its own absence.

Subsetting had no owner either: shaping columns for a 弥生/freee importer is industry-format work, which
CLAUDE.md places downstream, and a template could not deliver it anyway (headers come from DisplayName; it
cannot rename or reorder across types). #414's rationale for keeping the *download* is untouched — it argued
the download must exist for an operator with no AI client and no code, never that templates must.

Replaces POST /api/vault-extract/export-templates/export with POST /api/vault-extract/documents/export:

- Columns = the 4 fixed system columns, then every LIVE FieldDefinition of the type in DisplayOrder.
  Decision (a), signed off on #499: a value a document still holds for an ARCHIVED definition no longer
  appears. The template path traversed soft-delete to resolve columns it explicitly referenced; nothing
  references them now. Export_omits_an_archived_field_definition_even_when_documents_hold_its_values pins it
  — and goes red the moment the export re-disables the soft-delete filter.
- documentTypeCode is required (fields are type-scoped, mirroring the FieldFilters rule); an unknown type
  loud-fails rather than emitting a header-only CSV, which would be a silent lie about the layer's contents.
- File name is {typeCode}-{yyyyMMdd-HHmmss}.{csv|xlsx}: ASCII and sortable, where DisplayName is CJK.
- The DocumentIds branch went with the templates. It was unreachable from the only UI (abp-extensible-table
  has no row selection) and its "ignores every filter" semantics was the very screen/file divergence #496
  removed.

Also makes the column order total. IFieldDefinitionRepository.GetListAsync ordered by DisplayOrder alone, but
DisplayOrder defaults to 0, so ties were ordinary and broken by whatever the DB returned. Now (DisplayOrder,
Name) — Name is unique per (TenantId, DocumentTypeId). Without it the same data would export in a different
column order on different days, and would disagree with the list, which renders from the same call. Both
guards were verified by mutation: each one, when removed, reddens exactly its own test and nothing else.

Deleted: ExportTemplate / ExportColumn / ExportTemplateManager / IExportTemplateRepository /
EfCoreExportTemplateRepository / ExportTemplateAppService + its 4 DTOs + interface / ExportTemplateController /
the VaultExportTemplates table (migration V499_DropExportTemplates — destructive; Down() recreates the schema
but not the rows, and nothing references the table) / permission keys VaultExtract.Documents.Templates.* /
error codes Extract:ExportTemplate*. Kept and unchanged: ExportFormat, ExportFileBuilder, ExportProjection,
DocumentFieldQueryResolver, Extract:ExportDocumentLimitExceeded, the MaxExportDocumentCount fail-fast,
IMultiTenant isolation, and the Documents.Export authorization.

EF 128 / Application 345 / Domain 185 green; solution 0 errors. The Angular half follows.
