+++ core/src/Dignite.Vault.Extract.Application.Contracts/Documents/Exports/ExportDocumentsInput.cs
@@ -1,21 +1,37 @@
 using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
+using Dignite.Vault.Extract.Documents.DocumentTypes;
+using Volo.Abp.Validation;
 
 namespace Dignite.Vault.Extract.Documents.Exports;
 
 /// <summary>
-/// Export execution input. When <see cref="DocumentIds"/> is non-empty, exports by selected IDs and
-/// ignores all filter fields; otherwise matches current-layer documents by filters, subject to the
-/// per-export document count limit.
-/// Document type is not specified here: templates are always type-bound
-/// (<c>ExportTemplate.DocumentTypeId</c>), and exports are always narrowed to the template's type.
+/// Export execution input (#499). Matches current-layer documents of <see cref="DocumentTypeCode"/> by the same
+/// filters the operator document list composes, and emits the four fixed system columns followed by <b>every live
+/// field defined on that type</b>, ordered by <c>FieldDefinition.DisplayOrder</c>. Subject to the per-export
+/// document count limit, which fails fast rather than truncating.
+/// <para>
+/// There is no saved column projection and no explicit id set. #499 deleted the export-template layer because a
+/// template's only persisted content was "which of this type's fields, in what order" — and ordering is already
+/// owned by <c>DisplayOrder</c>, the same axis the list renders its columns by. One axis, so the file cannot
+/// disagree with the screen. The old <c>DocumentIds</c> branch went with it: it was unreachable from the only UI
+/// (<c>abp-extensible-table</c> has no row selection) and its "ignores every filter" semantics was exactly the
+/// screen/file divergence #496 set out to remove.
+/// </para>
 /// </summary>
 public class ExportDocumentsInput
 {
-    public Guid TemplateId { get; set; }
+    /// <summary>
+    /// Required: extracted fields are type-scoped, so an export must name exactly one type. Mirrors the existing
+    /// rule that <see cref="FieldFilters"/> requires a type on the list / MCP search path. A mixed-type view has
+    /// no field columns and cannot be exported.
+    /// </summary>
+    [Required]
+    [DynamicStringLength(typeof(DocumentTypeConsts), nameof(DocumentTypeConsts.MaxTypeCodeLength))]
+    public string DocumentTypeCode { get; set; } = default!;
 
-    public List<Guid>? DocumentIds { get; set; }
+    public ExportFormat Format { get; set; } = ExportFormat.Csv;
 
     public DocumentLifecycleStatus? LifecycleStatus { get; set; }
 
@@ -26,26 +42,23 @@ public class ExportDocumentsInput
     public DateTime? CreationTimeMax { get; set; }
 
     /// <summary>
-    /// #496: operator review-queue filter, the export twin of <see cref="GetDocumentListInput.HasReviewReasons"/>
-    /// (#284 / #395). <c>true</c> exports only documents that still require operator attention, evaluated by the
-    /// canonical <c>DocumentReviewQueries.RequiresAttention</c> predicate the document list and the needs-review
-    /// badge already run — never a second, hand-rolled copy. Ignored when <see cref="DocumentIds"/> is supplied.
+    /// Operator review-queue filter (#284 / #395), evaluated by the canonical
+    /// <c>DocumentReviewQueries.RequiresAttention</c> predicate the document list and the needs-review badge
+    /// already run — never a second, hand-rolled copy.
     /// </summary>
     public bool? HasReviewReasons { get; set; }
 
     /// <summary>
-    /// #496: provenance filter, the export twin of <see cref="GetDocumentListInput.OriginDocumentId"/> (#354).
-    /// When set, exports only the sub-documents derived from this source document (those whose
-    /// <c>Document.OriginDocumentId</c> equals it). Ignored when <see cref="DocumentIds"/> is supplied. Stays
-    /// under the ABP <c>IMultiTenant</c> global filter, so it can only ever reach the caller's own layer.
+    /// Provenance filter (#354): when set, exports only the sub-documents derived from this source document.
+    /// Stays under the ABP <c>IMultiTenant</c> global filter, so it can only ever reach the caller's own layer.
     /// </summary>
     public Guid? OriginDocumentId { get; set; }
 
     /// <summary>
-    /// #414: extracted-field-value filters, resolved against the template's document type and AND-combined
-    /// with the metadata filters above (reuses the same machinery as the document list / MCP search —
-    /// <c>ApplyFieldValueFilter</c> via <c>GetFieldMatchedIdsAsync</c>). Ignored when <see cref="DocumentIds"/>
-    /// is supplied. Each element self-validates; the list is capped like the list/search path.
+    /// Extracted-field-value filters, resolved against <see cref="DocumentTypeCode"/> and AND-combined with the
+    /// metadata filters above through the same machinery as the document list / MCP search
+    /// (<c>DocumentFieldQueryResolver</c> → <c>GetFieldMatchedIdsAsync</c>). Each element self-validates; the
+    /// list is capped like the list/search path.
     /// </summary>
     [MaxLength(DocumentConsts.MaxSearchFieldFilters)]
     public List<DocumentFieldFilter>? FieldFilters { get; set; }
