--- core/src/Dignite.Vault.Extract.Domain/Documents/Exports/IExportTemplateRepository.cs
@@ -1,12 +1,0 @@
-using System;
-using System.Threading;
-using System.Threading.Tasks;
-using Volo.Abp.Domain.Repositories;
-
-namespace Dignite.Vault.Extract.Documents.Exports;
-
-public interface IExportTemplateRepository : IRepository<ExportTemplate, Guid>
-{
-    /// <summary>Finds a template by current layer and Name, used for duplicate checks on creation.</summary>
-    Task<ExportTemplate?> FindByNameAsync(string name, CancellationToken cancellationToken = default);
-}
