+++ core/src/Dignite.Vault.Extract.EntityFrameworkCore/Documents/Fields/EfCoreFieldDefinitionRepository.cs
@@ -23,7 +23,14 @@ public class EfCoreFieldDefinitionRepository
         var dbSet = await GetDbSetAsync();
         return await dbSet
             .Where(f => f.DocumentTypeId == documentTypeId)
+            // #499: ThenBy(Name) makes this a total order. DisplayOrder defaults to 0, so ties are ordinary
+            // (a pack import, an admin who never set an order) and were previously broken by whatever the DB
+            // returned. Since the export now derives its columns from this very sequence, an unstable tail
+            // would mean the same data exports in a different column order on different days — and would
+            // disagree with the operator list, which renders from the same call. Name is unique per
+            // (TenantId, DocumentTypeId), so the order is total.
             .OrderBy(f => f.DisplayOrder)
+            .ThenBy(f => f.Name)
             .ToListAsync(GetCancellationToken(cancellationToken));
     }
